+++ src/util/binaryReader.h
@@ -6,24 +6,26 @@
 #define LOW_NIBBLE 0
 #define HIGH_NIBBLE 1
 
+#include "util/util.h"
+
 struct BinaryReader {
     char* buffer;
-    int pointer; // pointer to the current byte
-    int eob; // end of buffer; 0 false, 1 true
-    int nibble_flag;
-    size_t buffer_size;
+    uint32 pointer; // pointer to the current byte
+    uint8 eob; // end of buffer; 0 false, 1 true
+    uint8 nibble_flag;
+    uint32 buffer_size;
 };
 
-void CreateBinaryReader(struct BinaryReader* binaryReader, char* file, size_t buffer_size);
+void CreateBinaryReader(struct BinaryReader* binaryReader, char* file, uint32 buffer_size);
 
-void AdvancePointer(struct BinaryReader* binaryReader, int numberToAdvance);
+void AdvancePointer(struct BinaryReader* binaryReader, uint32 numberToAdvance);
 
 
 // reads current nibble and advances pointer
 char ReadNibble(struct BinaryReader* binaryReader);
 
 // reads the what_nibble at pointer where what_nibble = 0 is low, 1 is high;
-char ReadSpecifiedNibble(struct BinaryReader* binaryReader, unsigned int pointer, unsigned int what_nibble);
+char ReadSpecifiedNibble(struct BinaryReader* binaryReader, uint32 pointer, uint32 what_nibble);
 
 // reads byte
 char ReadByte(struct BinaryReader* binaryReader);
